Complete restructure, cleanup
