+++ client/src/components/dashboard/pages/Scan.js
